Implement merging rules to play 1024
This allows the user to swipe left, right, up and down to play 1024.

The iteration part focuses on building the order in which to go
through the tiles. To achieve this a matrix of coordinates is
constructed, which is then used to iterate through the tiles.

Building the order in which to iterate the tiles first alleviates
complexity by making the merging part agnostic to the order in which
it has to do that.

By doing so the merging operation can focus on how to handle:

1. Merging of tiles

    - Iterating through the coordinates in reverse order allows us to
      merge and place the coordinates in their proper positions at
      the same time.

2. Moving tiles into the proper positions

    - Which may or may not contain 2 tiles with multiple empty tiles
      in between, 2 adjacent tiles, 3 tiles with the same value, ...